+++ packages/core/navigation/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atlaskit/navigation
 
+## 32.3.3
+
+
+- [none] Updated dependencies [da63331](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/da63331)
+  - @atlaskit/button@8.2.5
+  - @atlaskit/item@7.0.8
+  - @atlaskit/quick-search@3.0.1
+  - @atlaskit/single-select@5.2.5
+  - @atlaskit/dropdown-menu@5.2.3
+  - @atlaskit/avatar@13.0.0
+- [patch] Updated dependencies [7724115](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7724115)
+  - @atlaskit/avatar@13.0.0
+  - @atlaskit/quick-search@3.0.1
+  - @atlaskit/button@8.2.5
+  - @atlaskit/single-select@5.2.5
+  - @atlaskit/item@7.0.8
+  - @atlaskit/dropdown-menu@5.2.3
+
 ## 32.3.2